epan: Widen loop variables in packet-[h-o]*
Fix

```
/builds/wireshark/wireshark/epan/dissectors/packet-hartip.c:952:18: warning: loop variable has narrower type 'int8_t' than iteration's upper bound 'uint8_t' [bugprone-too-small-loop-variable]
  952 |      for (i = 0; i < number_of_cmds; i++)
      |                  ^
```

and similar warnings.

+++ epan/dissectors/packet-oran.c
@@ -1303,7 +1303,7 @@ static void ext11_work_out_bundles(unsigned startPrbc,
         }
 
         unsigned bundles_set = 0;
-        for (uint8_t n=0;
+        for (unsigned n=0;
              n < (settings->ext6_num_bits_set * settings->ext6_rbg_size) / numBundPrb;
              n++) {
